Extra serde config support to a separate file
